libica: Bugfixes for asm, stck and programming samples
The loop in the asm used in the profiling, compiled with an earlier version
of gcc, but not with the actual. This had to be fixed. Also stckf
instruction is not available in 31 bit, thus in this case we have to use
stcke instead. Additionaly some fixes in the programming samples had to be
done.