From: Rainer Wolafka <rwolafka@de.ibm.com>
Problem description:
Libica uses different instructions to obtain the system clock when running in
31 and 64 bit. This information is used to generate entropy when initializing
the RPNG. The instructions take different input one of which is a buffer that
differs in size for the 31/64 instructions. When running in 31 bit this can
lead to a problem were memory is corrupted (writing beyond the buffer). This
problem does not occur when using libica in 64 bit mode.

Problem resolution:
Adjust the buffer size based on whether running in 31 or 64 bit mode.