Improve testcases for AES,3DES and DES stream ciphers.
Improvement on the ica testcases for the stream ciphers cfb, ofb and ctr.
Instead of just increasing the data size by blocksize with each iteration
now a random number between 1 and 8 is choosen. So the stream ciphers are
now tested with 'odd' data sizes - as they are all specified to work
on byte boundary. These improvements showed some issues with libica
and have been resolved with another fix.

Signed-off-by: Harald Freudenberger <hfreude@linux.vnet.ibm.com>