Fix aes and des/3des cfb mode with not blocksized input data.
In AES, DES and 3DES when input data was not a multiple of the
blocksize the unterlaying s390 assembler call may eventually
crash with 'illegal instruction'. Debugged, fixed and tested so
that the very last call may en/decrypt the whole blocksize
instead of the rest bytes but returns just the en/decrypted
rest bytes. This is correct as cfb is a stream cipher mode.

Signed-off-by: Harald Freudenberger <freude@linux.vnet.ibm.com>