icastats: Correct owner of shared segment when root creates it for a user
If a user calls in his session icatstats or any libica action then he creates
his shared segment for icastats counters. After this he can see his counts and
also if root looks with 'icastats --user' then we see the right counts.

But if root wants to see counts of a user (before this user did any actions
on libica and icastats) calling 'icastats --user' then the shared segment for
the user is created with wrong permissions and actions from the user are not
counted into this shared segment.

This patch ensures that the ownership of a shared segment created by root
for a user is changed to the user right after creation of the shared segment.

Signed-off-by: Ingo Franzki <ifranzki@linux.vnet.ibm.com>
Reviewed-by: Patrick Steuer <patrick.steuer@de.ibm.com>