Allow a zero output_length for ica_random_number_generate
In previous versions of libica, the ica_random_number_generate function
allowed to specify an output length of zero. The current implementation
returned -1 in this case when built with FIPS support. This causes errors
in some opencryptoki testcases (RSA PKCS PSS Sign).

This patch changes the behavior back to how it was in previous versions
of libica.

Signed-off-by: Ingo Franzki <ifranzki@linux.vnet.ibm.com>