Add libica fallback environment variable.
This patch introduces the LIBICA_FALLBACK_MODE environment
variable. During initialization of the Libica this variable
is checked and if valid the fallback mode is set accordingly.
The application can at any time later set the fallback mode
of Libica via ica_set_fallback_mode() and so "overwrite"
the setting done by the environment variable.

Signed-off-by: Harald Freudenberger <freude@linux.vnet.ibm.com>