Provide new compile option NO_CPACF and new libica-cex module
This compile option disables all functions based on CPACF. It is
used together with NO_SW_FALLBACKS to build an additional module
libica-cex.so.VERSION, which only provides hardware crypto support
based on IBM CryptoExpress adapters. The libica-cex module is intended
to be used in environments with special requirements in terms of
security and certification.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>