[PATCH 2/2] FIPS: Find libica from phdrs.
Instead of dlopening libica.so.3 by name, search through all phdrs and the
associated loadable program segment until the code for the search is found.
The binary associated with this phdr is the one that should be checked.

Add a sanity check to prevent multiple finds (impossible, but simple to check
for this kind of programming errors).  Fail if the library is either not found
at all or found multiple times.

This approach has the benefit of not needing a dlopen for a specific name and
can thus be used to also check libraries in non-standard locations.  This
might only be needed during development.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>