OpenSSL 3.0: Do not cleanup OpenSSL library context during library destructor
OpenSSL cleanup may have already run once the library destructor is called, this
may result in crashes. On the other hand, we can not register an OpenSSL cleanup
handler for this, because one may unload the library before OpenSSl cleanup runs,
this would also cause crashes.

So we can only not cleanup the library context at all, and leak it if one unloads
the library. OpenSSl will anyway clean up the contexts at program termination.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>