test: add test for adapter handle
Add a unit test for ica_open_adapter(). It is expected, that a process,
which calls this API call gets a file handle to the adapter, which should
not be greater than 3. If this test fails, the library initialization
opens a file descriptor, but not closing it in the end.

Signed-off-by: Holger Dengler <dengler@linux.ibm.com>