fips: allow to override the kernel fips indication
Allow libica tests in active fips mode on systems that are
not running in fips mode (i.e. the kernel fips flag is off).
New environment variable LIBICA_FIPS_FLAG=1 overrides the
kernel's fips indication and puts libica into active fips
mode. The fips_test is skipped in this case, because it
would fail due to this discrepancy.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>