FIPS 140-3: new API function ica_aes_gcm_initialize_fips
ica_aes_gcm_initialize_fips allows to create the GCM iv internally
via an approved random source and pass it back to the application
via the iv parm. So here the iv is an output parm and the application
must provide a writable buffer of sufficient length to receive the
internal iv. The minimum iv_length in fips mode is 12 bytes.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>