FIPS 140-3: new API function ica_get_hw_info
The new API function provides some hardware related info.
It is intended to be used in a fips environment together
with already existing function ica_get_msa_level to
determine the CPACF code level. Note that the msa level
alone is ambiguous, because it can be the same for multiple
machine generations.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>