Ensure OpenSSL config is loaded before creating libica's own library context
Ensure OpenSSL is initialized and the OpenSSL config is loaded BEFORE creating
libica's library context. Otherwise the OpenSSL config is loaded later, which
may cause that all configured providers (e.g. the IBMCA provider) are also
loaded into the library context. When the IBMCA provider is loaded into
libica's library context, then a recursion can happen, since the IBMCA provider
calls libica itself. Thus, we need to make sure that only the default or fips
provider is loaded in the library context.

This change is important for applications that do not itself initialize
OpenSSL. If they do, then libica's OpenSSL initialization is a no-op.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>