icastats: Ignore unknown user in summary and all report
If icastats is used with option '--summary' and it encounters a statistics
shared memory segment for a user that no longer exists, it fails with
'get_stats_sum: : Success'. This is because getpwuid() fails, but does not
set errno, so that perror prints 'Success'.
Instead of failing, just ignore such shared memory segments for users that
do no longer exist.

A similar problem exists with option '--all', it stops iterating over the
statistics shared memory segments if it finds one for a non existing user.
Continue the loop to return statistics for additional users.
Also fix a problem when stats_mmap() fails in the loop to close the directory
file descriptor before returning NULL.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>