ECC: fix KDSA sign behavior
If we are in fips mode, the k-value is created via openssl. This
may cause a cc=2 (random number not invertible) from the KDSA
instruction. In this case loop over KDSA with new random values.
But don't loop over calls to KDSA with unchanged k-value in the
non-fips case. This could lead to an endless loop.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>