Also check for deconfigured and check-stopped cards
When checking if a device is online, not only check the 'online' sysfs
attribute, but also check the 'config' and 'chkstop' attributes. Cards
and APQNs in check-stopped or deconfigured state can still be reported
as online via the sysfs attribute, although they are not available.
In case the 2 additional sysfs attributes are not available in sysfs,
then just rely on the 'online' attribute only.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>