fips update: Change service indicator implementation for AES GCM
Fix handling to differentiate if the call to AES-GCM encryption API was approved
or not. If the IV was set externally, it's non-approved, otherwise with internal
IV it's approved. Bind the service indicator to the service by checking the
behavior of the GCM IV in the gcm API.

Note that the GCM APIs are available in following cases:

1. Approved (either running in non-fips mode, or not using an external iv).
   In this case the GCM init functions are performed normally and return 0
   on successful completion.

2. Non-approved, but on override list (running in fips mode and using an
   external iv). In this case the GCM init functions are performed as well
   and also return 0 on successful completion, but the errno is set to EPERM
   to give applications the opportunity to distinguish the two cases.

Also note that the ica_aes_gcm API function, which does not need/use an init
function, is blocked unconditionally in fips mode already. So no handling
required for this API.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>