CONFIGURE: Make the OpenSSL FIPS config file name configurable
The name of the OpenSSL FIPS config file may be different on various
distros. It is included in src/openssl3-fips.cnf when used with
OpenSSL 3.0 or later.

To use a specific name:
  ./configure --enable-fips --with-fips-config=fips_local.cnf

The default remains fipsmodule.cnf. It is only used when --enable-fips
is also specified, and libica is built against OpenSSL 3.0 or later.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>